+++ NI_Fonakolos_játék/NI_Fonakolos_játék/NI_Fonakolos_játék.csproj.user
@@ -7,7 +7,15 @@
     </ApplicationDefinition>
   </ItemGroup>
   <ItemGroup>
+    <Compile Update="NewGameWindow.xaml.cs">
+      <SubType>Code</SubType>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
     <Page Update="MainWindow.xaml">
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Update="NewGameWindow.xaml">
       <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
