--- NI_Fonakolos_játék/NI_Fonakolos_játék/NewGame.xaml
@@ -1,28 +1,0 @@
-﻿<Window x:Class="NI_Fonakolos_játék.NewGameWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:NI_Fonakolos_játék"
-        mc:Ignorable="d"
-        Title="NewGameWindow" Height="294" Width="311" Background="#FF3B6439">
-
-
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="25*"/>
-            <ColumnDefinition Width="131*"/>
-        </Grid.ColumnDefinitions>
-        <Button Content="Computer VS Player" Tag="CvsP_cmp" HorizontalAlignment="Left" Margin="15,37,0,0" VerticalAlignment="Top" Width="120" Height="31" Click="CvsP_clicked" Grid.ColumnSpan="2"/>
-        <Button Content="Player VS Player" HorizontalAlignment="Left" Margin="15,90,0,0" VerticalAlignment="Top" Width="120" Height="32" Click="PvsP_clicked" Grid.ColumnSpan="2"/>
-        <Button Content="Exit Game" HorizontalAlignment="Left" Margin="132,238,0,0" VerticalAlignment="Top" Width="120" Height="30" Click="Exit_click" Grid.Column="1"/>
-        <TextBox  x:Name="Player1_box" HorizontalAlignment="Left" Height="29" Margin="106,139,0,0" TextWrapping="Wrap"  VerticalAlignment="Top" Width="145" Visibility="Visible" Grid.Column="1"/>
-        <TextBox  x:Name="Player2_box"  HorizontalAlignment="Left" Height="29" Margin="106,187,0,0" TextWrapping="Wrap"  VerticalAlignment="Top" Width="145" Visibility="Visible" Grid.Column="1"/>
-        <Button x:Name="playBTN" Content="PLAY" HorizontalAlignment="Left" Margin="15,238,0,0" VerticalAlignment="Top" Width="120" Height="30" Click="playBTN_Click" Grid.ColumnSpan="2"  />
-        <TextBlock x:Name="Player2"  Margin="15,187,145,0" TextWrapping="Wrap" VerticalAlignment="Top" FontSize="18" FontWeight="Bold" Visibility="Visible" Text="Second Player: " Grid.ColumnSpan="2"/>
-        <TextBlock x:Name="Player1"  Margin="15,144,185,0" TextWrapping="Wrap" VerticalAlignment="Top" FontSize="18" FontWeight="Bold" Visibility="Visible" Text="First Player: " Grid.ColumnSpan="2"/>
-
-
-
-    </Grid>
-</Window>
