--- NI_Fonakolos_játék/NI_Fonakolos_játék/NewGame.xaml.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-
-namespace NI_Fonakolos_játék
-{
-    /// <summary>
-    /// Interaction logic for NewGame.xaml
-    /// </summary>
-    public partial class NewGame : Window
-    {
-        public NewGame()
-        {
-            InitializeComponent();
-        }
-    }
-}
