+++ NI_Fonakolos_játék/NI_Fonakolos_játék/NewGameWindow.xaml.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace NI_Fonakolos_játék
+{
+    /// <summary>
+    /// Interaction logic for NewGameWindow.xaml
+    /// </summary>
+    public partial class NewGameWindow : Window
+    {
+        public NewGameWindow()
+        {
+            InitializeComponent();
+        }
+
+        private void CvsP_clicked(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+
+        private void playBTN_Click(object sender, RoutedEventArgs e)
+        {
+            MainWindow main = new MainWindow();
+            main.Show();
+            this.Close();
+        }
+
+        private void PvsP_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void exitBTN_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+    }
+}
